moved code into testing-module-assertion and testing-modules-load-testing
