+++ testing-modules-assertions/src/main/java/com/baeldung/jspec/Animal.java
