+++ testing-modules-assertions/src/main/java/com/baeldung/testing/assertj/Member.java
