+++ testing-modules-assertions/src/main/java/com/baeldung/testing/truth/User.java
