+++ testing-modules-assertions/src/test/java/com/testing/assertj/exceptions/Java7StyleAssertions.java
