+++ testing-modules-load-testing/gatling/src/test/scala/org/baeldung/RecordedSimulation.scala
