+++ testing-modules-load-testing/load-testing-comparison/src/main/java/com/baeldung/loadtesting/model/Transaction.java
