+++ testing-modules-load-testing/load-testing-comparison/src/main/resources/scripts/Gatling/GatlingScenario.scala
